move _seed_id_map helper function to inventory.util
